Mobile mixed feeds: per-item faction cards, not per-page chrome
Tasks + Praxis mobile feeds no longer theme the whole page to one faction.
Each card now picks its skin from its own data item's faction slug, mirroring
the desktop TaskCard/PraxisCard per-item dispatch.

Tasks: extracted each faction's bespoke mobile card (verbatim) from its
<Faction>TaskList page-skin into pages/tasks/mobileArchetypes/cards/, added a
MOBILE_TASK_CARD_BY_SLUG dispatcher (pickVariant on task.primary_faction_slug),
turned DefaultTasks into the single neutral shell, and deleted the 7 page-skins
+ their per-viewer dispatch tests (replaced by one dispatcher test).

Praxis: MobilePraxisFeed now renders the real PraxisCard / CollaborationCard per
item (interim per #565; bespoke mobile praxis cards are #573) and drops the
homemade MobileProofCard.

Relocated the 7 grandfathered TaskList raw-style exemptions to the 8 new card
files verbatim (same code, new paths — not new authoring; token migration of
these bespoke cards drains later per #579).

Closes #565
