+++ frontend/.eslint-legacy-raw-styles.txt
@@ -183,13 +183,14 @@ src/pages/taskDetail/mobileArchetypes/SnideTaskDetail.tsx
 src/pages/taskDetail/mobileArchetypes/UaTaskDetail.tsx
 src/pages/taskDetail/mobileArchetypes/WowTaskDetail.tsx
 src/pages/Tasks.tsx
-src/pages/tasks/mobileArchetypes/AlbescentTaskList.tsx
 src/pages/tasks/mobileArchetypes/DefaultTasks.tsx
-src/pages/tasks/mobileArchetypes/EphemeristsTaskList.tsx
-src/pages/tasks/mobileArchetypes/EverymenTaskList.tsx
-src/pages/tasks/mobileArchetypes/SingularityTaskList.tsx
-src/pages/tasks/mobileArchetypes/SnideTaskList.tsx
-src/pages/tasks/mobileArchetypes/UaTaskList.tsx
-src/pages/tasks/mobileArchetypes/WowTaskList.tsx
+src/pages/tasks/mobileArchetypes/cards/AlbescentMobileTaskCard.tsx
+src/pages/tasks/mobileArchetypes/cards/DefaultMobileTaskCard.tsx
+src/pages/tasks/mobileArchetypes/cards/EphemeristsMobileTaskCard.tsx
+src/pages/tasks/mobileArchetypes/cards/EverymenMobileTaskCard.tsx
+src/pages/tasks/mobileArchetypes/cards/SingularityMobileTaskCard.tsx
+src/pages/tasks/mobileArchetypes/cards/SnideMobileTaskCard.tsx
+src/pages/tasks/mobileArchetypes/cards/UaMobileTaskCard.tsx
+src/pages/tasks/mobileArchetypes/cards/WowMobileTaskCard.tsx
 src/pages/Updates.tsx
 src/pages/updates/MobileUpdates.tsx
