+++ frontend/src/pages/tasks/mobileArchetypes/DefaultTasks.tsx
@@ -1,16 +1,17 @@
-import { Link } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
-import type { TaskOut } from '../../../api/tasks'
 import { factionCssVar, factionName, sortFactionsByRainbowOrder } from '../../../utils/factions'
 import type { TasksState } from '../useTasks'
+import MobileTaskCard from './mobileTaskCard'
 
 /**
  * Default MOBILE task-browse skin — a scannable single-column card list with
  * phone-native filter chips (horizontal-scroll rows for status / faction /
  * level), NOT the desktop sidebar. Consumes the shared `useTasks()` state so a
  * chip tap updates the same filter state that keys the task read, and the
- * rendered set changes. Every faction falls through here until it registers a
- * bespoke mobile skin. Mirrors the DefaultTaskDetail mobile idiom (#496–#500).
+ * rendered set changes. The page chrome is faction-agnostic; each card in the
+ * results list picks its own skin from its task's faction slug via the
+ * `MobileTaskCard` dispatcher (#565). Mirrors the DefaultTaskDetail mobile idiom
+ * (#496–#500).
  */
 export default function DefaultTasks({ state }: { state: TasksState }) {
   const { t } = useTranslation('tasks')
@@ -165,79 +166,5 @@ function Chip({
       )}
       {children}
     </button>
-  )
-}
-
-function MobileTaskCard({ task, points }: { task: TaskOut; points: number }) {
-  const { t } = useTranslation('tasks')
-  const color = factionCssVar(task.primary_faction_slug)
-  return (
-    <Link
-      to={`/tasks/${task.id}`}
-      className="sidebar-card"
-      style={{
-        display: 'flex',
-        flexDirection: 'column',
-        gap: 8,
-        padding: '14px 16px 14px 18px',
-        borderLeft: `4px solid ${color}`,
-        textDecoration: 'none',
-      }}
-    >
-      {/* Faction tag */}
-      <span
-        className="eyebrow"
-        style={{ display: 'inline-flex', alignItems: 'center', gap: 6, color }}
-      >
-        <i style={{ width: 8, height: 8, borderRadius: 2, background: color, flex: 'none' }} />
-        {factionName(task.primary_faction_slug)}
-      </span>
-
-      {/* Title */}
-      <h2
-        className="font-display italic font-medium"
-        style={{ fontSize: 18, lineHeight: 1.15, color: 'var(--color-text-primary)' }}
-      >
-        {task.title}
-      </h2>
-
-      {/* Description — clamp to two lines */}
-      {task.description && (
-        <p
-          className="font-body"
-          style={{
-            fontSize: 13,
-            lineHeight: 1.5,
-            color: 'var(--color-text-secondary)',
-            display: '-webkit-box',
-            WebkitLineClamp: 2,
-            WebkitBoxOrient: 'vertical',
-            overflow: 'hidden',
-          }}
-        >
-          {task.description}
-        </p>
-      )}
-
-      {/* Footer — points + level */}
-      <div className="flex items-center gap-3" style={{ marginTop: 2 }}>
-        <span
-          style={{
-            fontFamily: "'Courier Prime', monospace",
-            fontSize: 12,
-            fontWeight: 700,
-            color: 'var(--color-text-primary)',
-            background: 'var(--color-bg-surface-alt)',
-            border: '1px solid var(--color-border)',
-            padding: '3px 9px',
-          }}
-        >
-          {t('mobile.points', { points })}
-        </span>
-        <span className="eyebrow" style={{ color: 'var(--color-text-tertiary)' }}>
-          {t('mobile.level', { level: task.level_required })}
-        </span>
-      </div>
-    </Link>
   )
 }
